Move Contexts to root
