+++ src/index.js
@@ -1,10 +1,10 @@
 import App from "./Admin";
 
+export { default as AdminContext, useAdmin } from "./AdminContext";
 export { default as Container } from "./Container";
 export { default as Content } from "./Content";
-export { default as AdminContext, useAdmin } from "./context/AdminContext";
-export { default as PageContext, usePage } from "./context/PageContext";
 export { default as Link } from "./Link";
+export { default as PageContext, usePage } from "./PageContext";
 export { default as LoginForm } from "./pages/LoginPageForm";
 export { default as TitleBar } from "./TitleBar";
 export { default as ToolBar } from "./ToolBar";
